Allow specifying what should be a fatal error.
Currently any nonzero status code received in the XML sent to
`receiveResponseXML` is considered a fatal error and the QBWC
session is torn down. That's incorrect according to the QBXML
documentation, which says that codes 1-499 are informational
and 500-999 are warnings. This adds a new configuration key,
`status_error_threshold`, which is the minimum status code
which should be considered a fatal error. To preserve reverse
compatibility the default is set to 1.